+++ example/integration_test/sc-IL-internalLimits/IL_200_key_length_truncate_test.dart
@@ -67,12 +67,14 @@ void main() {
       } else if (a == 5) {
         // 0) Custom Event
         var eventRaw = json.decode(queryParams['events']![0]);
-        Map<String, dynamic> event = eventRaw[0];
+        var number = Platform.isIOS ? 0 : 1; // orientation in android
+        Map<String, dynamic> event = eventRaw[number];
         expect(event['key'], 'Event With Sum And Segment'.substring(0, MAX_KEY_LENGTH));
         expect(event['segmentation']['Country'.substring(0, MAX_KEY_LENGTH)], 'Turkey');
         expect(event['segmentation']['Age'.substring(0, MAX_KEY_LENGTH)], '28884');
         // 1) View Start (legacy)
-        Map<String, dynamic> view = eventRaw[1];
+        number++;
+        Map<String, dynamic> view = eventRaw[number];
         expect(view['key'], '[CLY]_view');
         expect(view['segmentation']['Cats'.substring(0, MAX_KEY_LENGTH)], '12345');
         // merges with moose if limit is small
@@ -85,14 +87,16 @@ void main() {
         expect(view['segmentation']['NotCamel'.substring(0, MAX_KEY_LENGTH)], 'Deerz');
         expect(view['segmentation']['Moose'.substring(0, MAX_KEY_LENGTH)], 'Deer');
         // 2) View End (legacy)
-        view = eventRaw[2];
+        number++;
+        view = eventRaw[number];
         expect(view['key'], '[CLY]_view');
         expect(view['segmentation']['segment'], Platform.isIOS ? 'iOS' : 'Android');
         expect(view['segmentation']['name'], 'HomePage'.substring(0, MAX_KEY_LENGTH));
         expect(view['segmentation']['Camel'.substring(0, MAX_KEY_LENGTH)], 666);
         expect(view['segmentation']['NotCamel'.substring(0, MAX_KEY_LENGTH)], 'Deerz');
         // 3) View Start (AutoStopped)
-        view = eventRaw[3];
+        number++;
+        view = eventRaw[number];
         expect(view['key'], '[CLY]_view');
         expect(view['segmentation']['Cats'.substring(0, MAX_KEY_LENGTH)], 12345);
         // expect(view['segmentation']['Moons'.substring(0, MAX_KEY_LENGTH)], 9.9866); // merges with moose