+++ example/integration_test/utils.dart
@@ -29,13 +29,73 @@ Future<List<String>> getEventQueue() async {
 void testCommonRequestParams(Map<String, List<String>> requestObject) {
   expect(requestObject['app_key']?[0], APP_KEY);
   expect(requestObject['sdk_name']?[0], "dart-flutterbnp-${Platform.isIOS ? "ios" : "android"}");
-  expect(requestObject['sdk_version']?[0], '24.7.0');
+  expect(requestObject['sdk_version']?[0], '24.7.1');
   expect(requestObject['av']?[0], Platform.isIOS ? '0.0.1' : '1.0.0');
   assert(requestObject['timestamp']?[0] != null);
 
   expect(requestObject['hour']?[0], DateTime.now().hour.toString());
   expect(requestObject['dow']?[0], DateTime.now().weekday.toString());
   expect(requestObject['tz']?[0], DateTime.now().timeZoneOffset.inMinutes.toString());
+}
+
+/// Verify custom request queue parameters
+/// This method checks if the provided parameter key matches the parameter value in the request queue
+Future<void> testLastRequestParams(Map<String, dynamic> params) async {
+  print('params: $params');
+
+  // Get request and event queues from native side
+  final requestList = await getRequestQueue(); // List of strings
+
+  // Some logs for debugging
+  print('RQ: $requestList');
+  print('RQ length: ${requestList.length}');
+
+  // Verify the request queue for a single request
+  if (requestList.length > 0) {
+    final queryParams = Uri.parse('?' + requestList.last).queryParametersAll;
+    print('queryParams: $queryParams');
+    testCommonRequestParams(queryParams);
+    for (final param in params.keys) {
+      expect(queryParams[param]?[0], params[param]);
+    }
+  } else {
+    if (params.isNotEmpty) {
+      // test failed.
+      expect(requestList.length, 'Test failed because request queue should not be empty');
+    }
+  }
+}
+
+/// Verify custom request queue parameters
+/// This method checks if the provided parameter key matches the parameter value in the request queue
+Future<String> testDeviceID(dynamic deviceIDMatcher) async {
+  // Get the device ID
+  String? id = await Countly.getCurrentDeviceId();
+  String? newModuleId = await Countly.instance.deviceId.getID();
+  // Verify the device ID
+  expect(id, deviceIDMatcher);
+  expect(newModuleId, deviceIDMatcher);
+  expect(id, newModuleId);
+  return id!;
+}
+
+/// Verify custom request queue parameters
+/// This method checks if the provided parameter key matches the parameter value in the request queue
+Future<DeviceIdType> testDeviceIDType(DeviceIdType givenType) async {
+  // Get the device ID type
+  DeviceIdType? type = await Countly.getDeviceIDType();
+  DeviceIdType? newModuleType = await Countly.instance.deviceId.getIDType();
+  // Verify the device ID type
+  expect(type, givenType);
+  expect(newModuleType, givenType);
+  if (givenType == DeviceIdType.SDK_GENERATED) {
+    String? id = await Countly.getCurrentDeviceId();
+    String? newModuleId = await Countly.instance.deviceId.getID();
+    expect(id!.length, Platform.isIOS ? 36 : 16);
+    expect(newModuleId!.length, Platform.isIOS ? 36 : 16);
+    expect(id, newModuleId);
+  }
+  return type!;
 }
 
 /// Start a server to receive the requests from the SDK and store them in a provided List
@@ -344,4 +404,37 @@ Future<void> testConsentForRC({bool isAT = false, bool isCG = true, bool isCNR =
     }
     await getAndValidateAllRecordedRCValues();
   }
+}
+
+// For session tests:
+void checkBeginSession(Map<String, List<String>> queryParams, {String deviceID = ''}) {
+  expect(queryParams['begin_session']?[0], '1');
+  if (deviceID.isNotEmpty) {
+    expect(queryParams['device_id']?[0], deviceID);
+  }
+}
+
+void checkMerge(Map<String, List<String>> queryParams, {String deviceID = '', String oldDeviceID = ''}) {
+  expect(queryParams['old_device_id']?[0].isNotEmpty, true);
+  if (deviceID.isNotEmpty) {
+    expect(queryParams['device_id']?[0], deviceID);
+  }
+  if (oldDeviceID.isNotEmpty) {
+    expect(queryParams['old_device_id']?[0], oldDeviceID);
+  }
+}
+
+void checkEndSession(Map<String, List<String>> queryParams, {String deviceID = ''}) {
+  expect(queryParams['end_session']?[0].isNotEmpty, true);
+//   expect(queryParams['session_duration']?[0].isNotEmpty, true); TODO: check this
+  if (deviceID.isNotEmpty) {
+    expect(queryParams['device_id']?[0], deviceID);
+  }
+}
+
+void printQueues(List<String> requestList, List<String> eventList) {
+  print('RQ: $requestList');
+  print('RQ length: ${requestList.length}');
+  print('EQ: $eventList');
+  print('EQ length: ${eventList.length}');
 }